+++ src/Shoelace.Style/Events/PlayChangeEventArgs.cs
@@ -1,0 +1,8 @@
+﻿namespace Shoelace.Style.Events;
+
+/// <inheritdoc/>
+public class PlayChangeEventArgs : EventArgs
+{
+    /// <inheritdoc/>
+    public bool Play { get; set; }
+}